+++ src/main/kotlin/com/react/musicServer/controllers/ApiController.kt
@@ -1,9 +1,10 @@
 package com.react.musicServer.controllers
 
 import com.react.musicServer.data.Data
-import com.react.musicServer.data.MessageData
+import com.react.musicServer.data.message.MessageData
+import com.react.musicServer.data.filepart.RemoteFilePart
 import com.react.musicServer.services.MainService
-import com.react.musicServer.services.YtUploadService
+import com.react.musicServer.services.YoutubeUploadService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.MediaType
 import org.springframework.http.codec.multipart.FilePart
@@ -17,7 +18,7 @@ import java.util.*
 @RequestMapping("api")
 class ApiController @Autowired constructor(
     private val service: MainService,
-    private val ytUploadService: YtUploadService,
+    private val youtubeUploadService: YoutubeUploadService,
 ) {
 //    @GetMapping("download/{uuid:.+}")
 //    suspend fun download(@PathVariable uuid: UUID): ByteArray {
@@ -41,25 +42,25 @@ class ApiController @Autowired constructor(
         val result = service.upload(RemoteFilePart(file))
         return MessageData(
             fileName = result.filename,
-            fileUrl = "/api/download/${result.id}",
-            fileUUID = result.id,
+            fileUrl = "/api/download/${result.uuid}",
+            fileUUID = result.uuid,
             fileTranscoded = result.wasTranscoded,
             message = "File uploaded with success!"
         )
     }
 
     @RequestMapping(
-        value = ["upload/yt"],
+        value = ["upload/youtube"],
         method = [RequestMethod.POST],
         consumes = [MediaType.APPLICATION_FORM_URLENCODED_VALUE]
     )
     suspend fun uploadFromYoutube(exchange: ServerWebExchange): MessageData {
         val formData = exchange.awaitFormData()
-        val result = ytUploadService.upload(formData["url"]?.firstOrNull()!!)
+        val result = youtubeUploadService.upload(formData["url"]?.firstOrNull()!!)
         return MessageData(
             fileName = result.filename,
-            fileUrl = "/api/download/${result.id}",
-            fileUUID = result.id,
+            fileUrl = "/api/download/${result.uuid}",
+            fileUUID = result.uuid,
             fileTranscoded = result.wasTranscoded,
             message = "Audio uploaded with success!"
         )
