+++ src/main/kotlin/com/react/musicServer/data/filepart/RemoteFilePart.kt
@@ -1,6 +1,5 @@
-package com.react.musicServer.controllers
+package com.react.musicServer.data.filepart
 
-import com.react.musicServer.data.RemoteFile
 import kotlinx.coroutines.reactor.awaitSingleOrNull
 import org.springframework.http.codec.multipart.FilePart
 import java.nio.file.Path
