+++ src/main/kotlin/com/react/musicServer/data/message/UploadResult.kt
@@ -1,0 +1,7 @@
+package com.react.musicServer.data.message
+
+data class UploadResult(
+    val uuid: String,
+    val filename: String,
+    val wasTranscoded: Boolean,
+)
