+++ src/main/kotlin/com/react/musicServer/exceptions/DownloadFailedException.kt
@@ -1,0 +1,3 @@
+package com.react.musicServer.exceptions
+
+class DownloadFailedException(message: String) : YoutubeUploadException(message)
