+++ src/main/kotlin/com/react/musicServer/exceptions/InvalidLinkException.kt
@@ -1,0 +1,3 @@
+package com.react.musicServer.exceptions
+
+class InvalidLinkException(message: String) : YoutubeUploadException(message)
