+++ src/main/kotlin/com/react/musicServer/exceptions/YoutubeUploadException.kt
@@ -1,0 +1,3 @@
+package com.react.musicServer.exceptions
+
+sealed class YoutubeUploadException(override val message: String) : RuntimeException(message)
