--- src/main/kotlin/com/react/musicServer/services/exceptions/DownloadFailedException.kt
@@ -1,3 +1,0 @@
-package com.react.musicServer.services.exceptions
-
-class DownloadFailedException(message: String) : YtUploadException(message)
