--- src/main/kotlin/com/react/musicServer/services/exceptions/InvalidLinkException.kt
@@ -1,3 +1,0 @@
-package com.react.musicServer.services.exceptions
-
-class InvalidLinkException(message: String) : YtUploadException(message)
