--- src/main/kotlin/com/react/musicServer/services/exceptions/UploadSizeExceededException.kt
@@ -1,3 +1,0 @@
-package com.react.musicServer.services.exceptions
-
-class UploadSizeExceededException(message: String) : RuntimeException(message)
