--- src/main/kotlin/com/react/musicServer/services/exceptions/YtUploadException.kt
@@ -1,3 +1,0 @@
-package com.react.musicServer.services.exceptions
-
-sealed class YtUploadException(override val message: String) : RuntimeException(message)
